# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloud-Ready School Management ERP API - Phase 1",
        "description": "Complete RESTful API specifications for Multi-Campus School Management ERP. Encompasses Institute Management, Multi-Campus Branches, Academic Years & Terms, Spatie Role-Based Access Control (RBAC), Global Settings, Classes, Sections, Streams, Houses, Subjects, Employee & Staff Management, and Student Admissions & Profiles.",
        "version": "1.0.0",
        "contact": {
            "name": "School ERP Support",
            "email": "support@schoolerp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Laravel Development API Server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Sanctum API Bearer Token"
            }
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Users (Paginated & Filterable)",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            },
            "post": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create User",
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get User Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details"
                    }
                }
            },
            "put": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted"
                    }
                }
            }
        },
        "/users/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "User & Role Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Toggle User Active/Inactive Status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User status toggled"
                    }
                }
            }
        },
...
        "/subjects": {
            "get": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "List Subjects",
                "responses": {
                    "200": {
                        "description": "List of subjects"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Create Subject",
                "responses": {
                    "201": {
                        "description": "Subject created"
                    }
                }
            }
        },
        "/subjects/{id}": {
            "get": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Get Subject Details",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Update Subject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Delete Subject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject deleted"
                    }
                }
            }
        },
        "/classes/{classId}/assign-subject": {
            "post": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Assign Subject to Class",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject mapped to class"
                    }
                }
            }
        },
        "/classes/{classId}/subjects/{subjectId}": {
            "delete": {
                "tags": [
                    "Subject Management"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Remove Subject from Class",
                "parameters": [
                    {
                        "name": "classId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "subjectId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject removed"
                    }
                }
            }
        },
...
    }
}